requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# requirements.txt
# ... (тут могут быть другие пакеты, оставь их)
pandas
pymongo
apache-airflow-providers-mongo
